Stage4_Enterprise_Practice/Lesson07_Document_Review_Agent/AgentTeamProject/backend/pyproject.toml
[project]
name = "contract-review-backend"
version = "0.1.0"
description = "合同审核系统后端 MVP"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "sqlalchemy>=2.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "python-multipart>=0.0.9",
    "python-dotenv>=1.0.0",
    "langchain>=0.3.0",
    "langchain-openai>=0.2.0",
    "langgraph>=0.2.0",
    "langgraph-checkpoint-sqlite>=2.0.0",
    "pypdf2>=3.0.0",
    "python-docx>=1.1.0",
    "sse-starlette>=2.1.0",
    "aiofiles>=24.1.0",
    "httpx>=0.27.0",
    "fpdf2>=2.8.7",
]

[tool.uv]
dev-dependencies = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.27.0",
]
